Update styles and layout of admin interface
Text inputs which set a certain property now can show the placeholder
text for that property. This means that we do not need to add a label
in the html page, so it keeps the interface compact.

The collapsible buttons now have more margin.

The network editing portion is now collapsible.

Instead of having the title of 'Stage', the stage collapsible components
show a summary of what's inside the stage. This is actually useful
information, as opposed to the 'Stage'.